Don't panic (on general_section for empty ini)
When creating a new ini or loading it from an input which contains no
actual entries but empty lines for the general section, calling
general_section and general_section_mut panics. This is not documented
nor do I expect this behavior: Isn't an empty file equivalent to an
empty general section? And from this perspective, calls to
general_section or general_section_mut should not panic.