fix: CI on a supported platform
Travis CI is no longer a platform this project can build on, so I moved it to CircleCI
where we can leverage the matrix feature for multiple platforms.

Running the CI jobs on multiple platforms revealed a bug in a test where Windows was
not using \r\n line endings.